# api/openapi.yaml
openapi: 3.1.0
info:
  title: Farmish API
  version: 0.1.0
  summary: Farmish Ghana agricultural marketplace API.
  description: |
    Contract-first HTTP API for Farmish. This file is the source of truth:
    Go server types are generated from it (`make generate`), and requests are
    validated against it at runtime.

    Conventions:
    - JSON properties and query parameters are camelCase.
    - Money is always integer pesewas (`Money.amount`) plus `currency: GHS`.
    - Errors use one envelope: `{"error": {"code", "message", "details?"}}`.
      `code` is a stable snake_case identifier; `message` is human-readable
      and may change.
    - List endpoints take `page` (from 1) and `limit` (1-50) and return `PageMeta`.
    - Authenticated endpoints take a Firebase ID token as `Authorization: Bearer <token>`.
      Every operation requires it unless it declares `security: []`. Missing,
      expired or invalid tokens all get the same 401 `unauthorized` envelope.
    - `x-farmish-role: admin` on an operation additionally requires the admin
      role (403 `forbidden` otherwise). The server enforces exactly what the
      spec declares.
    - `x-farmish-step-up: true` on an operation requires a fresh sign-in
      (Firebase `auth_time` within 5 minutes) verified with a revocation
      check. A stale sign-in gets 401 `reauth_required` with
      `WWW-Authenticate: Bearer error="insufficient_user_authentication"`,
...
      tokens still get the generic 401 `unauthorized`.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/me/seller-profile:
    get:
      tags: [sellers]
      operationId: getMySellerProfile
      summary: Get my seller profile
      description: Returns the signed-in user's seller profile, or 404 when they never created one.
      responses:
        "200":
          description: The caller's seller profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    put:
      tags: [sellers]
      operationId: updateMySellerProfile
      summary: Create or edit my seller profile
      description: |
        Creates the profile on first call, edits it afterwards. Submitting
        or changing idType/idNumber encrypts the number, stores its last 4
        and (re)submits the profile for verification. Editing other fields
        never changes the verification status.
      x-farmish-rate-limit: sensitive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellerProfileInput"
      responses:
        "200":
          description: The saved seller profile. Never contains the ID number.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/sellers/{userId}:
    get:
      tags: [sellers]
      operationId: getPublicSeller
      summary: Get a seller's public profile
      description: Safe projection only, without contact, identity or account fields.
      security: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The seller's public profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicSeller"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/admin/sellers:
    get:
      tags: [sellers]
      operationId: listAdminSellers
      summary: List seller profiles by verification status
      description: Admin review queue, oldest submission first.
      x-farmish-role: admin
      parameters:
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum: [unverified, pending, verified, rejected]
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of seller profiles with owner contact fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfileAdminList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/admin/sellers/{userId}/verification:
    post:
      tags: [sellers]
      operationId: verifySeller
      summary: Approve or reject a seller
      description: |
        Decides a pending profile. Approve verifies it; reject needs a
        reason. Anything not pending gets 409.
      x-farmish-role: admin
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifySellerRequest"
      responses:
        "200":
          description: The decided profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfileAdmin"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token.
  parameters:
...
    Page:
      name: page
      in: query
      description: 1-based page number.
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Page size.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 50
        default: 20
  responses:
    BadRequest:
      description: The request is malformed or fails validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, expired or invalid credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated, but not allowed to do this.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The resource is in a state that forbids this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TooManyRequests:
      description: Rate limit exceeded. Retry after `Retry-After` seconds.
      headers:
        Retry-After:
          description: Seconds until a request will be allowed.
          schema:
            type: integer
        X-RateLimit-Limit:
          $ref: "#/components/headers/X-RateLimit-Limit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/X-RateLimit-Remaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/X-RateLimit-Reset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  headers:
...
    X-RateLimit-Limit:
      description: Burst capacity of the limit that applied.
      schema:
        type: integer
    X-RateLimit-Remaining:
      description: Requests left before limiting.
      schema:
        type: integer
    X-RateLimit-Reset:
      description: Seconds until the limit is fully replenished.
      schema:
        type: integer
  schemas:
...
    Error:
      type: object
      description: The single error envelope used by every endpoint.
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            Stable snake_case error code: `bad_request`, `validation_failed`,
            `unauthorized`, `reauth_required`, `forbidden`, `not_found`,
            `method_not_allowed`, `rate_limited`, `turnstile_failed`,
            `invalid_transition`, `seller_profile_required`,
            `listing_suspended`, `internal_error`, `unavailable`.
            `reauth_required` means the sign-in is too old for a step-up
            operation; `invalid_transition` and `listing_suspended` are
            state conflicts (409).
          examples: [validation_failed]
        message:
          type: string
          description: Human-readable summary. Do not parse.
          examples: [Request validation failed]
        details:
          type: array
          description: Per-field problems, present for `validation_failed`.
          items:
            $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [location, message]
      properties:
        location:
          type: string
          enum: [path, query, header, cookie, body]
        field:
          type: string
          description: Parameter name, or a JSON pointer into the body (e.g. `/items/0/qty`).
          examples: [/items/0/qty]
        message:
          type: string
          examples: [must be >= 1]
...
    SellerProfileInput:
      type: object
      description: Create-or-edit model for the caller's seller profile. idType and idNumber must be given together.
      additionalProperties: false
      required: [businessName, region, district]
      properties:
        businessName:
          type: string
          minLength: 2
          maxLength: 120
        region:
          type: string
          enum: [Ahafo, Ashanti, Bono, Bono East, Central, Eastern, Greater Accra, North East, Northern, Oti, Savannah, Upper East, Upper West, Volta, Western, Western North]
        district:
          type: string
          minLength: 2
          maxLength: 80
        bio:
          type: string
          maxLength: 1000
        showPhone:
          type: boolean
        showWhatsapp:
          type: boolean
        whatsapp:
          type: string
          description: Raw input; normalised server-side to E.164 (+233 followed by 9 digits).
        idType:
          type: string
          enum: [ghana_card, passport, voters_id, drivers_license]
        idNumber:
          type: string
          minLength: 4
          maxLength: 30
          pattern: '^[A-Za-z0-9-]+$'
    SellerProfile:
      type: object
      description: The owner's view of their seller profile. Never contains the ID number, only its last 4.
      required: [businessName, region, district, showPhone, showWhatsapp, verificationStatus]
      properties:
        businessName:
          type: string
        region:
          type: string
        district:
          type: string
        bio:
          type: string
        showPhone:
          type: boolean
        showWhatsapp:
          type: boolean
        whatsapp:
          type: string
          description: E.164 number, present when the seller set one.
        verificationStatus:
          type: string
          enum: [unverified, pending, verified, rejected]
        idType:
          type: string
          enum: [ghana_card, passport, voters_id, drivers_license]
        idNumberLast4:
          type: string
          description: Masked ID number (last 4) for the owner and admins.
        submittedAt:
          type: string
          format: date-time
        reviewedAt:
          type: string
          format: date-time
        rejectionReason:
          type: string
...
    PublicSeller:
      type: object
      description: Safe seller projection for anonymous callers.
      required: [userId, businessName, region, district, verified, memberSince, rating]
      properties:
        userId:
          type: string
          format: uuid
        businessName:
          type: string
        region:
          type: string
        district:
          type: string
        bio:
          type: string
        verified:
          type: boolean
          description: True when the seller is verified.
        memberSince:
          type: string
          format: date-time
          description: When the seller's user account was created.
        rating:
          $ref: "#/components/schemas/SellerRating"
    SellerProfileAdmin:
      type: object
      description: The admin review view, with the owner's contact fields.
      required: [userId, businessName, region, district, showPhone, showWhatsapp, verificationStatus]
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
        email:
          type: string
        phone:
          type: string
        businessName:
          type: string
        region:
          type: string
        district:
          type: string
        bio:
          type: string
        showPhone:
          type: boolean
        showWhatsapp:
          type: boolean
        whatsapp:
          type: string
        verificationStatus:
          type: string
          enum: [unverified, pending, verified, rejected]
        idType:
          type: string
          enum: [ghana_card, passport, voters_id, drivers_license]
        idNumberLast4:
          type: string
        submittedAt:
          type: string
          format: date-time
        reviewedAt:
          type: string
          format: date-time
        rejectionReason:
          type: string
    SellerProfileAdminList:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SellerProfileAdmin"
        meta:
          $ref: "#/components/schemas/PageMeta"
    VerifySellerRequest:
      type: object
      additionalProperties: false
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approve, reject]
        reason:
          type: string
          maxLength: 500
          description: Required when rejecting (1-500 characters).
...
    PageMeta:
      type: object
      required: [page, limit, total]
      properties:
        page:
          type: integer
          format: int32
          minimum: 1
        limit:
          type: integer
          format: int32
          minimum: 1
          maximum: 50
        total:
          type: integer
          format: int64
          minimum: 0
          description: Total matching items across all pages.
...
    SellerRating:
      type: object
      description: A seller aggregate over visible reviews.
      required: [average, count]
      properties:
        average:
          type: number
          description: Mean rating to one decimal; 0 with no reviews.
          examples: [4.5]
        count:
          type: integer
          format: int64
          minimum: 0
...